group cache structure to single file
