--- src/main/kotlin/org/bibletranslationtools/fetcher/usecase/cache/BookCache.kt
@@ -1,8 +1,0 @@
-package org.bibletranslationtools.fetcher.usecase.cache
-
-class BookCache(
-    val slug: String,
-    var availability: Boolean = false,
-    val url: String? = null,
-    val chapters: List<ChapterCache> = listOf()
-)
