--- src/main/kotlin/org/bibletranslationtools/fetcher/usecase/cache/ChapterCache.kt
@@ -1,7 +1,0 @@
-package org.bibletranslationtools.fetcher.usecase.cache
-
-data class ChapterCache(
-    val number: Int,
-    var availability: Boolean = false,
-    var url: String? = null
-)
