--- src/main/kotlin/org/bibletranslationtools/fetcher/usecase/cache/LanguageCache.kt
@@ -1,7 +1,0 @@
-package org.bibletranslationtools.fetcher.usecase.cache
-
-class LanguageCache(
-    val code: String,
-    var availability: Boolean = false,
-    val products: List<ProductCache> = listOf()
-)
