--- src/main/kotlin/org/bibletranslationtools/fetcher/usecase/cache/ProductCache.kt
@@ -1,7 +1,0 @@
-package org.bibletranslationtools.fetcher.usecase.cache
-
-class ProductCache(
-    val slug: String,
-    var availability: Boolean = false,
-    val books: List<BookCache> = listOf()
-)
